requirements.txt
dlib==19.18.0
face_recognition
matplotlib
Pillow